Move Zendesk reconciler code to pkg/reconciler and cmd/
