--- zendesk/Makefile
@@ -1,9 +1,0 @@
-KSOURCE      = zendesk
-KSOURCE_DESC = Triggermesh Zendesk Event Source for Knative
-BASE_DIR    := $(CURDIR)/..
-
-include $(BASE_DIR)/scripts/inc.Makefile
-
-GOPKGS = ./cmd/... ./pkg/reconciler/...
-
-include $(BASE_DIR)/scripts/inc.Target
