--- zendesk/config/201-clusterroles.yaml
@@ -1,126 +1,0 @@
-# Copyright (c) 2020 TriggerMesh Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: zendesk-source-controller
-rules:
-- apiGroups:
-  - serving.knative.dev
-  resources:
-  - services
-  verbs: &everything
-  - get
-  - list
-  - watch
-  - create
-  - update
-  - patch
-  - delete
-
-# Record Kubernetes events
-- apiGroups:
-  - ''
-  resources:
-  - events
-  verbs:
-  - create
-  - update
-  - patch
-
-  # Read secrets
-- apiGroups:
-  - ''
-  resources:
-  - secrets
-  verbs:
-  - get
-
-- apiGroups:
-  - sources.triggermesh.io
-  resources:
-  - zendesksources
-  verbs:
-  - list
-  - watch
-  - get
-
-- apiGroups:
-  - sources.triggermesh.io
-  resources:
-  - zendesksources/status
-  verbs:
-  - update
-
-# Read controller configurations
-- apiGroups:
-  - ''
-  resources:
-  - configmaps
-  verbs: &listwatch
-  - list
-  - watch
-- apiGroups:
-  - ''
-  resources:
-  - configmaps
-  resourceNames:
-  - config-logging
-  - config-observability
-  - config-leader-election
-  verbs:
-  - get
-
-# Resolve sink URIs
-- apiGroups:
-  - ''
-  resources:
-  - services
-  verbs: *listwatch
-- apiGroups:
-  - serving.knative.dev
-  resources:
-  - services
-  verbs: *listwatch
-- apiGroups:
-  - eventing.knative.dev
-  resources:
-  - brokers
-  verbs: *listwatch
-- apiGroups:
-  - messaging.knative.dev
-  resources:
-  - channels
-  verbs: *listwatch
-
----
-
-# The role is needed for the aggregated role source-observer in knative-eventing to provide readonly access to "Sources".
-# See https://github.com/knative/eventing/blob/master/config/200-source-observer-clusterrole.yaml.
-kind: ClusterRole
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: zendesk-source-observer
-  labels:
-    duck.knative.dev/source: "true"
-rules:
-- apiGroups:
-  - sources.triggermesh.io
-  resources:
-  - zendesksources
-  verbs:
-  - get
-  - list
-  - watch
