--- zendesk/sample/400-source.yaml
@@ -1,35 +1,0 @@
-# Copyright (c) 2020 TriggerMesh Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: sources.triggermesh.io/v1alpha1
-kind: ZendeskSource
-metadata:
-  name: zendesksource
-spec:
-  email: '<ZENDESK ACCOUNT USER EMAIL>'
-  subdomain: '<ZENDESK TENANT SUBDOMAIN>'
-  token:
-    secretKeyRef:
-      name: zendesksource
-      key: token
-  webhookUsername: '<SOME ARBITRARY USERNAME>'
-  webhookPassword:
-    secretKeyRef:
-      name: zendesksource
-      key: password
-  sink:
-    ref:
-      apiVersion: serving.knative.dev/v1
-      kind: Service
-      name: event-display
